remove duplicate dialog and port to Wizard
